Move stuff into users module
