--- client/src/main/res/layout/content_users_list.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.swiperefreshlayout.widget.SwipeRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                              xmlns:app="http://schemas.android.com/apk/res-auto"
-                                              xmlns:tools="http://schemas.android.com/tools"
-                                              android:id="@+id/users_refresh_swipe_layout"
-                                              android:layout_width="match_parent"
-                                              android:layout_height="match_parent"
-                                              android:paddingBottom="@dimen/activity_vertical_margin"
-                                              android:paddingLeft="@dimen/activity_horizontal_margin"
-                                              android:paddingRight="@dimen/activity_horizontal_margin"
-                                              app:layout_behavior="@string/appbar_scrolling_view_behavior"
-                                              tools:context=".ui.UsersActivity"
-                                              tools:showIn="@layout/activity_users_list">
-
-  <androidx.recyclerview.widget.RecyclerView
-    android:id="@+id/users_recycler"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"/>
-</androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
