+++ feature-users/src/main/java/com/jraska/github/client/users/UsersModule.kt
@@ -1,16 +1,21 @@
 package com.jraska.github.client.users
 
+import android.app.Activity
 import androidx.lifecycle.ViewModel
 import com.jraska.github.client.Config
 import com.jraska.github.client.Navigator
 import com.jraska.github.client.PerApp
 import com.jraska.github.client.analytics.EventAnalytics
+import com.jraska.github.client.core.android.LinkLauncher
 import com.jraska.github.client.rx.AppSchedulers
+import com.jraska.github.client.users.ui.UsersActivity
 
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.ClassKey
 import dagger.multibindings.IntoMap
+import dagger.multibindings.IntoSet
+import okhttp3.HttpUrl
 import retrofit2.Retrofit
 
 @Module
@@ -63,5 +68,32 @@ object UsersModule {
     analytics: EventAnalytics
   ): ViewModel {
     return RepoDetailViewModel(repository, schedulers, navigator, analytics)
+  }
+
+  @JvmStatic
+  @Provides
+  @IntoSet
+  fun provideUsersPathLauncher(): LinkLauncher {
+    return UsersPathLauncher()
+  }
+
+  @JvmStatic
+  @Provides
+  @IntoSet
+  fun provideUsersListLauncher(): LinkLauncher {
+    return object : LinkLauncher {
+      override fun launch(inActivity: Activity, deepLink: HttpUrl): LinkLauncher.Result {
+        return if ("/users" == deepLink.encodedPath()) {
+          UsersActivity.start(inActivity)
+          LinkLauncher.Result.LAUNCHED
+        } else {
+          LinkLauncher.Result.NOT_LAUNCHED
+        }
+      }
+
+      override fun priority(): LinkLauncher.Priority {
+        return LinkLauncher.Priority.EXACT_MATCH
+      }
+    }
   }
 }
