+++ feature-users/src/main/java/com/jraska/github/client/users/ui/RepoDetailActivity.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.users.ui
 
 import android.app.Activity
 import android.content.Intent
@@ -7,13 +7,14 @@ import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.airbnb.epoxy.SimpleEpoxyAdapter
 import com.airbnb.epoxy.SimpleEpoxyModel
-import com.jraska.github.client.R
 import com.jraska.github.client.core.android.BaseActivity
+import com.jraska.github.client.core.android.viewModel
+import com.jraska.github.client.users.R
 import com.jraska.github.client.users.RepoDetail
 import com.jraska.github.client.users.RepoDetailViewModel
-import com.jraska.github.client.core.android.viewModel
-import kotlinx.android.synthetic.main.activity_repo_detail.*
-import kotlinx.android.synthetic.main.content_repo_detail.*
+import kotlinx.android.synthetic.main.activity_repo_detail.repo_detail_github_fab
+import kotlinx.android.synthetic.main.activity_repo_detail.toolbar
+import kotlinx.android.synthetic.main.content_repo_detail.repo_detail_recycler
 
 class RepoDetailActivity : BaseActivity() {
