+++ feature-users/src/main/java/com/jraska/github/client/users/ui/ReposSectionModel.kt
@@ -1,9 +1,9 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.users.ui
 
 import android.view.View
 import com.airbnb.epoxy.EpoxyModel
 import com.airbnb.epoxy.SimpleEpoxyAdapter
-import com.jraska.github.client.R
+import com.jraska.github.client.users.R
 import com.jraska.github.client.users.RepoHeader
 import kotlinx.android.synthetic.main.item_repos_section.view.*
