+++ feature-users/src/main/java/com/jraska/github/client/users/widget/RepeaterLayout.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.widget
+package com.jraska.github.client.users.widget
 
 import android.annotation.TargetApi
 import android.content.Context
