+++ feature-users/src/main/res/drawable/ic_admin_24dp.xml
