+++ feature-users/src/main/res/drawable/ic_followers_48dp.xml
