+++ feature-users/src/main/res/drawable/ic_following_48dp.xml
