+++ feature-users/src/main/res/drawable/ic_github_48dp.xml
