+++ feature-users/src/main/res/drawable/ic_history_24dp.xml
