+++ feature-users/src/main/res/drawable/ic_settings_white_24dp.xml
