+++ feature-users/src/main/res/drawable/ic_signal_wifi_off_black_48dp.xml
