+++ feature-users/src/main/res/drawable/ic_source_fork_24dp.xml
