+++ feature-users/src/main/res/drawable/ic_star_black_24dp.xml
