+++ feature-users/src/main/res/drawable/ic_star_black_48dp.xml
