+++ feature-users/src/main/res/layout/activity_repo_detail.xml
@@ -5,22 +5,30 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.UsersActivity">
+    tools:context=".ui.RepoDetailActivity">
 
   <com.google.android.material.appbar.AppBarLayout
       android:layout_height="wrap_content"
       android:layout_width="match_parent"
-      android:theme="@style/AppTheme.AppBarOverlay">
+      android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
         android:background="?attr/colorPrimary"
-        app:popupTheme="@style/AppTheme.PopupOverlay" />
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
   </com.google.android.material.appbar.AppBarLayout>
 
-  <include layout="@layout/content_users_list" />
+  <include layout="@layout/content_repo_detail" />
+
+  <com.google.android.material.floatingactionbutton.FloatingActionButton
+    android:id="@+id/repo_detail_github_fab"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_margin="@dimen/fab_margin"
+    android:layout_gravity="bottom|end"
+    android:src="@drawable/ic_github_48dp" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
