+++ feature-users/src/main/res/layout/activity_user_detail.xml
@@ -5,14 +5,14 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.jraska.github.client.ui.UserDetailActivity">
+    tools:context=".ui.UserDetailActivity">
 
   <com.google.android.material.appbar.AppBarLayout
       android:id="@+id/app_bar"
       android:fitsSystemWindows="true"
       android:layout_height="@dimen/app_bar_height"
       android:layout_width="match_parent"
-      android:theme="@style/AppTheme.AppBarOverlay">
+      android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
     <com.google.android.material.appbar.CollapsingToolbarLayout
         android:id="@+id/toolbar_layout"
@@ -35,7 +35,7 @@
           android:layout_height="?attr/actionBarSize"
           android:layout_width="match_parent"
           app:layout_collapseMode="pin"
-          app:popupTheme="@style/AppTheme.PopupOverlay" />
+          app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
     </com.google.android.material.appbar.CollapsingToolbarLayout>
   </com.google.android.material.appbar.AppBarLayout>
