+++ feature-users/src/main/res/layout/activity_users_list.xml
@@ -10,25 +10,17 @@
   <com.google.android.material.appbar.AppBarLayout
       android:layout_height="wrap_content"
       android:layout_width="match_parent"
-      android:theme="@style/AppTheme.AppBarOverlay">
+      android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
         android:background="?attr/colorPrimary"
-        app:popupTheme="@style/AppTheme.PopupOverlay" />
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
   </com.google.android.material.appbar.AppBarLayout>
 
-  <include layout="@layout/content_repo_detail" />
-
-  <com.google.android.material.floatingactionbutton.FloatingActionButton
-    android:id="@+id/repo_detail_github_fab"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_margin="@dimen/fab_margin"
-    android:layout_gravity="bottom|end"
-    android:src="@drawable/ic_github_48dp" />
+  <include layout="@layout/content_users_list" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
