+++ feature-users/src/main/res/layout/content_user_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.jraska.github.client.ui.UserDetailActivity"
+    tools:context=".ui.UserDetailActivity"
     tools:showIn="@layout/activity_user_detail">
 
   <androidx.recyclerview.widget.RecyclerView
