+++ feature-users/src/main/res/layout/item_loading.xml
