+++ feature-users/src/main/res/layout/item_repo_detail_stats.xml
