+++ feature-users/src/main/res/layout/item_simple_text.xml
