+++ feature-users/src/main/res/layout/item_user_stats.xml
