+++ feature-users/src/main/res/values/dimens.xml
@@ -1,15 +1,16 @@
 <resources>
+  <dimen name="swipe_progress_circle_diameter">40dp</dimen>
+
+  <dimen name="fab_margin">16dp</dimen>
+
   <dimen name="side_padding">16dp</dimen>
   <dimen name="icon_size">48dp</dimen>
   <dimen name="padding_small">4dp</dimen>
   <dimen name="padding_medium">8dp</dimen>
 
-  <dimen name="fab_margin">16dp</dimen>
   <dimen name="app_bar_height">180dp</dimen>
   <dimen name="text_margin">16dp</dimen>
 
   <dimen name="elevation_resting">2dp</dimen>
   <dimen name="corner_radius">3dp</dimen>
-
-  <dimen name="swipe_progress_circle_diameter">40dp</dimen>
 </resources>
