+++ feature-users/src/main/res/values/strings.xml
@@ -1,0 +1,11 @@
+<resources>
+  <string name="user_detail_joined_template">  Joined on %s</string>
+  <string name="repos_popular">Popular repositories</string>
+  <string name="repos_contributed">Contributed repositories</string>
+  <string name="repo_detail_language_used_template">Language: %s</string>
+  <string name="repo_detail_issues_template">Issues Count: %s</string>
+
+  <string name="settings">Settings</string>
+  <string name="github_clickable_icon">GitHub clickable icon</string>
+  <string name="administrator_icon">Administrator icon</string>
+</resources>
